% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumo.R
\name{addStops}
\alias{addStops}
\title{addStops}
\description{
For every pedestrian route walking along a road, if he/she travels to a junction connected to
a building, then this function will augment the route he/she is on to travel inside the building, waiting for
a specific duration (designated by the building type), and then walks out and conintues on the original route
}
