% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumo.R
\name{getPolygons}
\alias{getPolygons}
\title{Get SUMO Polygons}
\description{
Retrieves Polygons in SUMO network
}
\details{
#' @title Get SUMO Edges
#' @name getEdges
#' @description Retrieves edges in SUMO road network
sumo$methods(getEdges = function(simplify = FALSE){
  sumo_net <- getNetwork()
  if (simplify){
    res <- as_list(xml_find_all(sumo_net, "//edge"))
    res <- lapply(res, function(edge) append(attributes(edge), list(lanes=attributes(edge$lane))))
    return(res)
  } else {
    return(xml2::xml_find_all(sumo_net, xpath = "//edge"))
  }
})
}
