% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumo.R
\docType{class}
\name{sumo}
\alias{sumo}
\title{SUMO Reference Class}
\usage{
## Make sure SUMO_HOME is set to SUMO's locations,
## i.e. Sys.setenv(SUMO_HOME=...)
sumo_net <- sumo$new()
sumo_net$<method name> ...
}
\value{
An reference class object of class "sumo". Methods are accessed with the standard '$'.
}
\description{
Reference class implementation that enables easy API-access to SUMO functionality suitable for simulation. A few
methods have been added which partially incorporate SUMO's Traffic Control Interface (TraCI) for fast retrieval of certain
types of simulation queries. Additional commands have been added which parse through SUMO's various input files (such as the
SUMO road network itself).
}
\details{
Before creating a SUMO object, there are two environmental variables to be aware of. The SUMO_HOME variable must be
set in the environmental prior to creating an object, otherwise it will return an error. Once set, SUMO requires several files
to be saved to the disk prior to creating a simulation. Namely, the SUMO road network must be set up and configured with either
a preconfigured road network, or via some external source (i.e. OSM) with \code{netconvert}. By default, these files are saved
in the 'default' directory and can be accessed via \code{system.file(..., package = "sumor")}. Alternatively, if the SUMO_TMP
variable is defined in the environment, all files will be saved relative to there.
}

\section{Methods}{

\subsection{Network and Configuration}{
  \itemize{
    \item{\strong{\code{\link{netconvert}}:}}{ Covert a given source to a SUMO road network.}
    \item{\strong{\code{\link{getOSM}}:}}{ Download an OpenStreetMap tile source.}
    \item{\strong{\code{\link{getNetwork}}:}}{ Retrieve the SUMO road network.}
    \item{\strong{\code{\link{getJunctions}}:}}{ Retrieve the junctions in the SUMO road network.}
    \item{\strong{\code{\link{getPolygons}}:}}{ Retrieve the polygons in the SUMO road network.}
    \item{\strong{\code{\link{addPolys}}:}}{ Add polygon information to the SUMO road network.}
    \item{\strong{\code{\link{mergeTrips}}:}}{ Merge two or more trips files into one.}
    \item{\strong{\code{\link{getEdges}}:}}{ Get the edges of the SUMO road network.}
    \item{\strong{\code{\link{getPedestrianEdges}}:}}{ Retrieve pedestrian-only edges from the SUMO road network. }
    \item{\strong{\code{\link{getTargetNodes}}:}}{ Retrieve the junctions in the SUMO road network, by edge id.}
  }
}
\subsection{Simulation}{
  \itemize{
    \item{\strong{\code{\link{simulate}}:}}{ Run the configured SUMO simulation.}
    \item{\strong{\code{\link{quickSim}}:}}{ Creates a quick SUMO simulation with suitable defaults.}
    \item{\strong{\code{\link{viewSimulation}}:}}{ View the simulation with the SUMO GUI.}
    \item{\strong{\code{\link{randomTrips}}:}}{ Creates random trips suitable for simulation.}
    \item{\strong{\code{\link{getRoutes}}:}}{ Retrieve the routes in the current simulation.}
    \item{\strong{\code{\link{getVehiclePositions}}:}}{ Retrieve the vehicle positions in the current simulation.}
    \item{\strong{\code{\link{genConfig}}:}}{ Generate configuration suitable for simulation.}
    \item{\strong{\code{\link{vehicleProbe}}:}}{ Probe a vehicle for various information.}
    \item{\strong{\code{\link{vehicleAcc}}:}}{ Retrieve vehicle acceleration.}
    \item{\strong{\code{\link{getFCD}}:}}{ Retrieve vehicle 'Floating Car Data.'}
  }
}
\subsection{Utility}{
  \itemize{
    \item{\strong{\code{\link{command}}:}}{ Runs a SUMO command.}
    \item{\strong{\code{\link{projectCRS}}:}}{ Projects SUMO coordinates to arbitrary CRS.}
    \item{\strong{\code{\link{toMeters}}:}}{ Convert WGS84 coordinates to Cartesian coordinates.}
    \item{\strong{\code{\link{plotOSM}}:}}{ View the region to use for the simulation with leaflet.}
 }
}
}

\examples{
library("sumor"); library("sp"); library("lubridate")

## Select a bounding box; I use http://boundingbox.klokantech.com/
osu_bbox <- bbox(matrix(c(-83.021676,39.9933,-83.005166,40.004366), nrow = 2, byrow = TRUE))

## Create the sumo object
osu_net <- sumo$new()

## Run quick road simulation w/ default settings
osu_net$quickSim(bbox = osu_bbox, sim_len = lubridate::dhours(0.5), p = 4, n = 1) # on average generate 1 vehicle every 4 seconds

## View the simulation.
osu_net$viewSimulation()
}
\seealso{
\code{\link{ReferenceClasses}}
}
\author{
Matt Piekenbrock
}
