% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumor.R
\docType{package}
\name{sumor}
\alias{sumor}
\alias{sumor-package}
\title{sumor: A port of the SUMO API for R.}
\description{
sumor enables easy access and use of a subset of the Simulation for Urban MObility (SUMO) API.
}
\section{methods}{
NA
}

